refactor: post-3.6a consistency tidy in components/wizard
Three non-behavioural items. The refine seams were split between verb-first (step-opening's refineOpening / refineDescription) and noun-first (step-world's genreRefine / toneRefine / settingRefine); 3.6b was expected to pick one and did not, and 3.6b's cast step added no refine prop, so the tie stood. Standardised on verb-first, which is what all five shipped defaults already use (refineGenreAssist and siblings).

blank() was defined twice in the folder; lore-list.tsx now imports the step-world-logic copy it already imports four other helpers from. assist-list-logic.ts is renamed ai-assist-logic.ts to restore the folder's <component>-logic.ts pairing — it belongs to ai-assist.tsx, not to a component called assist-list.
